Event issues, bug fixes

+++ MixerFixerV1/00_Services/00_DB/DB_AudioObject.cs
@@ -15,7 +15,7 @@ namespace Services
         public bool IsActive { get; set; } = false;
         public bool IsMute { get; set; } = false;
         public bool IsDevice { get; set; } = false;
-        public double Volume { get; set; }
+        public int Volume { get; set; }
     }
 
     public partial class Srv_DB
